fix:Correccion de validacion

+++ src/main/java/com/petwellness/service/impl/HorariosDisponiblesServiceImpl.java
@@ -1,5 +1,8 @@
 package com.petwellness.service.impl;
 
+import com.petwellness.dto.HorariosDisponiblesDTO;
+import com.petwellness.exception.ResourceNotFoundException;
+import com.petwellness.mapper.HorariosDisponiblesMapper;
 import com.petwellness.model.entity.HorariosDisponibles;
 import com.petwellness.model.entity.Veterinario;
 import com.petwellness.repository.HorariosDisponiblesRepository;
@@ -7,6 +10,7 @@ import com.petwellness.repository.VeterinarioRepository;
 import com.petwellness.service.HorariosDisponiblesService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -15,30 +19,46 @@ import java.util.List;
 public class HorariosDisponiblesServiceImpl implements HorariosDisponiblesService {
     private final HorariosDisponiblesRepository horariosDisponiblesRepository;
     private final VeterinarioRepository veterinarioRepository;
+    private final HorariosDisponiblesMapper horariosDisponiblesMapper;
 
+    @Transactional
     @Override
-    public HorariosDisponibles agregarHorario(HorariosDisponibles horario) {
-        Veterinario veterinario = veterinarioRepository.findById(horario.getVeterinario().getUsuario_user_id())
-                .orElseThrow(() -> new RuntimeException("Veterinario no encontrado con id: " + horario.getVeterinario().getUsuario_user_id()));
+    public HorariosDisponiblesDTO agregarHorario(HorariosDisponiblesDTO horarioDTO) {
+        Veterinario veterinario = veterinarioRepository.findById(horarioDTO.getVeterinarioId())
+                .orElseThrow(() -> new ResourceNotFoundException("Veterinario no encontrado con id: " + horarioDTO.getVeterinarioId()));
 
+        HorariosDisponibles horario = horariosDisponiblesMapper.toEntity(horarioDTO);
         horario.setVeterinario(veterinario);
 
-        return horariosDisponiblesRepository.save(horario);
+        HorariosDisponibles nuevoHorario = horariosDisponiblesRepository.save(horario);
+
+        return horariosDisponiblesMapper.toDTO(nuevoHorario);
     }
 
+    @Transactional(readOnly = true)
+    @Override
+    public List<HorariosDisponiblesDTO> obtenerHorarios() {
+        List<HorariosDisponibles> horarios = horariosDisponiblesRepository.findAll();
+        return horarios.stream()
+                .map(horariosDisponiblesMapper::toDTO)
+                .toList();
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<HorariosDisponiblesDTO> obtenerHorariosPorVeterinarioId(Integer userId) {
+        List<HorariosDisponibles> horarios = horariosDisponiblesRepository.findByVeterinarioUsuarioUserId(userId);
+        return horarios.stream()
+                .map(horariosDisponiblesMapper::toDTO)
+                .toList();
+    }
+
+    @Transactional
     @Override
     public void eliminarHorario(Integer id) {
+        if (!horariosDisponiblesRepository.existsById(id)) {
+            throw new ResourceNotFoundException("El horario disponible no existe con id: " + id);
+        }
         horariosDisponiblesRepository.deleteById(id);
-
-    }
-
-    @Override
-    public List<HorariosDisponibles> obtenerHorarios() {
-        return horariosDisponiblesRepository.findAll();
-    }
-
-    @Override
-    public List<HorariosDisponibles> obtenerHorariosPorVeterinarioId(Integer userId) {
-        return horariosDisponiblesRepository.findByVeterinarioUsuarioUserId(userId);
     }
 }
